backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.4
django-allauth==65.3.0
django-cors-headers==4.6.0
django-vite==3.0.5
djangorestframework==3.15.2
idna==3.10
jwt==1.3.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.2.3
